+++ emitter/part1.md
@@ -28,6 +28,6 @@ For example:
 For convenience, here are some unit tests for our `EventEmitter` in order to clarify its behavior.
 
 ### running tests
-Name your file `emitter-part-1.js` and export your class/constructor as `EventEmitter`. After you `npm i` you can execute `npm run test`.
+Name your file `emitter.js` and export your class/constructor as `EventEmitter`. After you `npm i` you can execute `npm run test`.
 
 **we __encourage__ to run the tests**
